refactor(runtime): export generateCallback and rename to runEventHandler
